feat: allow entity detail to have a different base URL from the entity list (#500)
