+++ packages/data-explorer-ui/src/config/entities.ts
@@ -126,6 +126,7 @@ export interface DataSourceConfig {
   defaultParams?: {
     catalog: string;
   };
+  entityURL?: string;
   url: string;
 }
